+++ src/extension.ts
@@ -12,9 +12,10 @@ import { OxCompletionItemProvider } from './OxCompletion';
 import CodeManager from './OxCodeManager';
 import GoDefinitionProvider from './oxGoToDefininition';
 import { GenDoc } from './OxGenDoc';
-import { getExtensionCommands, IsCorrectOxFile, FixPathWindows } from './util';
-import { CheckOxMetricsIsOk, GetOxMetricsPath, GetOxDocFolder } from './OxBin';
+import { getExtensionCommands, IsCorrectOxFile, FixPathWindows, DevLog } from './util';
+import { CheckOxMetricsIsOk, GetOxMetricsPath, GetOxDocFolder, GetOxIncludeFolders, IsWindows } from './OxBin';
 import { OxDocumentFormattingEditProvider } from './OxFormat';
+import { OxIncludeCompletionItemProvider } from './OxIncludeCompletion';
 const OX_MODE: vscode.DocumentFilter = { language: 'ox', scheme: 'file' };
 const OX_SELECTOR: vscode.DocumentSelector = { scheme: 'file', language: 'ox' };// only files from disk
 const codeManager = new CodeManager();
@@ -23,19 +24,22 @@ const open = require('open');
 export function activate(context: vscode.ExtensionContext) {
 
     console.log("Ox for Visual Studio - version : ", extensionPackage["version"]);
-    console.log("vscode version : ", vscode.version);
-    console.log("process.platform : ", process.platform);
+    DevLog("vscode version : ", vscode.version);
+    DevLog("process.platform : ", process.platform);
     if (process.platform != "darwin" && process.platform != "win32" && process.platform != "linux") {
-        console.log("invalid platform");
+        DevLog("invalid platform");
         vscode.window.showInformationMessage("OxCode does not support the current platform");
         return;
     }
 
     if (!CheckOxMetricsIsOk(context.extensionPath)) {
-        console.log('Invalid configuration')
+        DevLog('Invalid configuration')
         return;
     }
     const SymbolProvider = new OxDocumentSymbolProvider();
+
+    const providerInclude = new OxIncludeCompletionItemProvider(GetOxIncludeFolders());
+    context.subscriptions.push(vscode.languages.registerCompletionItemProvider(OX_SELECTOR, providerInclude, "<", '"', "/", "\\"));
     context.subscriptions.push(vscode.languages.registerDocumentSymbolProvider(OX_MODE, SymbolProvider));
     context.subscriptions.push(vscode.languages.registerImplementationProvider(OX_SELECTOR, new GoImplementationProvider()));
     context.subscriptions.push(vscode.languages.registerDocumentFormattingEditProvider(OX_SELECTOR, new OxDocumentFormattingEditProvider()));
@@ -57,6 +61,7 @@ export function activate(context: vscode.ExtensionContext) {
     context.subscriptions.push(workspace.onDidChangeTextDocument(eve => GenDoc(eve)));
     context.subscriptions.push(workspace.onDidSaveTextDocument(document => CheckSyntaxOnSave(document)));
     context.subscriptions.push(workspace.onDidSaveTextDocument(() => codeManager.clearError()));
+    context.subscriptions.push(vscode.commands.registerCommand('extension.ox.GetDebuggerPath', config => { return GetDebuggerPath(); }));
     context.subscriptions.push(vscode.commands.registerCommand('ox.show.commands', () => {
         const extCommands = getExtensionCommands();
         extCommands.push({ command: 'editor.action.goToDeclaration', title: 'Go to Definition' });
@@ -70,9 +75,20 @@ export function activate(context: vscode.ExtensionContext) {
         });
     }));
 
-    console.log('OK activate extension [789]')
+    DevLog('OK activate extension [789]')
 }
+function GetDebuggerPath(): string {
 
+    if (IsWindows()) {
+        var oxConfig = vscode.workspace.getConfiguration('oxcode');
+        if (oxConfig == null)
+            return "tochange";
+        var oxfolder = oxConfig["oxmetricsFolder"];
+        var path = oxfolder + "\\ox\\bin\\oxli.exe";
+        return path;
+    }
+    return "tochange";
+}
 function OpenWithOxMetrics(): void {
     try {
         if (!IsCorrectOxFile())
@@ -89,11 +105,11 @@ function OpenWithOxMetrics(): void {
         });
         subprocess.unref();
     } catch (error) {
-        console.log(error);
+        DevLog(error);
     }
 }
 async function RunHelp(): Promise<void> {
-    console.log("RunHelp");
+    DevLog("RunHelp");
     if (!IsCorrectOxFile())
         return;
     try {
@@ -102,7 +118,7 @@ async function RunHelp(): Promise<void> {
         open(url2); // for some reasons it didn't work with vscode.env.openExterna(...)
 
     } catch (error) {
-        console.log(error);
+        DevLog(error);
     }
 }
 function CheckSyntaxOnSave(document: vscode.TextDocument): void {
@@ -117,7 +133,7 @@ function CleanBakFiles(): void {
     vscode.workspace.findFiles("*.bak", "", 20).then(value => {
         for (let _i = 0; _i < value.length; _i++) {
             var mess = value[_i];
-            console.log("Delete .bak file: " + mess);
+            DevLog("Delete .bak file: " + mess);
             fs.unlinkSync(mess.fsPath);
 
         }
@@ -125,7 +141,7 @@ function CleanBakFiles(): void {
 }
 // this method is called when your extension is deactivated
 export function deactivate() {
-    console.log("deactivate oxcode");
+    DevLog("deactivate oxcode");
     if (codeManager != null)
         codeManager.KillRunningOx();
 }